+++ src/rv/rv.service.ts
@@ -175,10 +175,14 @@ export class RvService {
         throw new NotFoundException(`RV with ID ${id} not found`)
       }
 
-      // Delete existing rv_items
-      await this.prisma.rVItem.deleteMany({
-        where: { rv_id: id },
-      });
+      if(input.items && input.items.length > 0){
+
+        // Delete existing meqs_items
+        await this.prisma.rVItem.deleteMany({
+          where: { rv_id: id },
+        });
+        
+      }
 
       const updatedRV = await this.prisma.rV.update({
         where: { id },
@@ -189,15 +193,15 @@ export class RvService {
           work_order_no: input.work_order_no ?? existingRV.work_order_no,
           work_order_date: input.work_order_date ? new Date(input.work_order_date) : existingRV.work_order_date,
           status: input.status ?? existingRV.status,
-          ...(input.classification_id
-            ? { classification: { connect: { id: input.classification_id } } }
-            : { classification: { connect: { id: existingRV.classification_id } } }),
-          ...(input.requested_by_id
-            ? { requested_by: { connect: { id: input.requested_by_id } } }
-            : { requested_by: { connect: { id: existingRV.requested_by_id } } }),
-          ...(input.supervisor_id
-            ? { supervisor: { connect: { id: input.supervisor_id } } }
-            : { supervisor: { connect: { id: existingRV.supervisor_id } } }),
+          classification: input.classification_id
+            ? { connect: { id: input.classification_id } }
+            : { connect: { id: existingRV.classification_id } },
+          requested_by: input.requested_by_id
+            ? { connect: { id: input.requested_by_id } }
+            : { connect: { id: existingRV.requested_by_id } },
+          supervisor: input.supervisor_id
+            ? { connect: { id: input.supervisor_id } }
+            : { connect: { id: existingRV.supervisor_id } },
           rv_items: {
             create: input.items?.map((item) => ({
               item: {
@@ -213,6 +217,40 @@ export class RvService {
         },
         include: { rv_items: { include: { item: true } } },
       });
+
+      // const updatedRV = await this.prisma.rV.update({
+      //   where: { id },
+      //   data: {
+      //     date_requested: input.date_requested ? new Date(input.date_requested) : existingRV.date_requested,
+      //     purpose: input.purpose ?? existingRV.purpose,
+      //     notes: input.notes ?? existingRV.notes,
+      //     work_order_no: input.work_order_no ?? existingRV.work_order_no,
+      //     work_order_date: input.work_order_date ? new Date(input.work_order_date) : existingRV.work_order_date,
+      //     status: input.status ?? existingRV.status,
+      //     ...(input.classification_id
+      //       ? { classification: { connect: { id: input.classification_id } } }
+      //       : { classification: { connect: { id: existingRV.classification_id } } }),
+      //     ...(input.requested_by_id
+      //       ? { requested_by: { connect: { id: input.requested_by_id } } }
+      //       : { requested_by: { connect: { id: existingRV.requested_by_id } } }),
+      //     ...(input.supervisor_id
+      //       ? { supervisor: { connect: { id: input.supervisor_id } } }
+      //       : { supervisor: { connect: { id: existingRV.supervisor_id } } }),
+      //     rv_items: {
+      //       create: input.items?.map((item) => ({
+      //         item: {
+      //           create: {
+      //             description: item.description,
+      //             brand: { connect: { id: item.brand_id } },
+      //             unit: { connect: { id: item.unit_id } },
+      //             quantity: item.quantity,
+      //           },
+      //         },
+      //       })),
+      //     },
+      //   },
+      //   include: { rv_items: { include: { item: true } } },
+      // });
       
 
       await this.prisma.$executeRaw`COMMIT`;
